Uninlined task
